window_launcher: close stdin on launch (fix #9)
A recent change in the game engine means that the launcher and client
processes now communicate via stdin, this seems to be causing issues
to do with them sharing a stdin fd with the CEF browser process.

If affected by this, a temporary workaround is to launch the game with
`</dev/null`.