browser: working main-thread messaging
So, CEF is very deficient in this area, both in its features and its
documentation. Lots of functions such as CefQuitMessageLoop can only be
called from the browser process main thread, but how do you send a
message to that thread from outside? The intended way is the message
pump flow, but that's totally unusable. What I've done instead is this:

- at entry point, spawn a browser and unmapped window
- check OnAfterCreated for that browser, if match, do entry point
- instead of running Exit(), destroy that browser window
- check OnBeforeClose for that browser, if match, run Exit()
- to message the main thread, call SendProcessMessage on that browser
  with TID_RENDERER
- in the render process handler, bounce the message back to TID_BROWSER
- the message is finally received in OnProcessMessageReceived in the
  browser process main thread

I despise it, but it's the only possible way.