client: fix crash on closing clients
Previously the correct client would me destroyed but it wasn't
necessarily the correct one that would get removed from the list. It
seems to be that it was always the last one that would get removed. I'm
not sure why this was using a multiple-result search in the first place.