window_launcher: use a default request handler
Browser::Launcher now intercepts all requests (except ones which fail to
parse). The default behaviour is very similar to how CEF would behave if
the request wasn't intercepted, except it reconstructs the CefRequest
without associating it with any browser or frame, meaning it's not
subject to cross-origin checks, and can present its own cross-origin
policy in its response to the webpage. This will finally put a stop to
the plague of issues with Jgx's login API having incorrect access
control settings.