browser: make Browser::Window self-managing
This was a long-overdue change to make Browser::Window be its own
CefClient and CefLifeSpanHandler, instead of making the main client
object receive callbacks for everything in the tree and try to delegate
it out. This should make everything much more robust, and also gets rid
of the DoClose override, which is "unsupported" (read: broken) in CEF
versions 126.x and onwards.