library: separate plugin API from backend
This is a very big change. It goes some way to breaking up the massive
wall of code that was plugin.c, and also removes all the documentation
from plugin_api.h, in preparation for it soon being properly documented
using Texinfo. It also changes the way function tables and metatables
are loaded to make it easier to add new ones (and reduce binary size a
lot.)