library: fix how embedded window positioning works
Now, if an embedded window or browser is forcibly repositioned by the
game window changing size, it will try to snap back to its previous
position and size, or as close as possible, when the game window is
resized again. The position it's trying to snap to will be overwritten
if the user repositions it again during this process.