library: revert 55a0185
The new engine version has been reverted, I'll reapply this patch when
(or if) the new version is re-released.

It's annoying that this situation is causing me to have to release new
versions on an almost daily basis, but I don't think it will be a
frequent occurrence. The plugin loader has been stable for over a year
up until this point.