app: completely redo functions.ts
Removed all XHR from this file, did a bit of dedupe, and removed a few
instances of what I assume must have been copy-paste errors. As a
by-product, one of the two remaining entries in store.ts was moved into
functions.ts, and the capitalisation of some fields of the Bolt
interface needed to be changed.