library: handle deleted texture bindings
Ugh.

For the most part, the game is pretty good at not causing use-after-free
situations in OpenGL terms. However there is one quite insidious case
where it'll bind a texture to unit 0, delete the texture, then try to
upload data with glCompressedTexSubImage2D. This is *definitely*
unintentional, and probably indicates that some kind of texturing glitch
is about to happen, but it's technically not an error in itself
according to the OpenGL docs:

> If a texture that is currently bound is deleted, the binding reverts
> to 0 (the default texture).

So this needs to be handled correctly, specifically by texture units
being IDs rather than pointers so that their validity can be checked
before each use.