app: move PSAs higher in hierarchy
They need to be outside of the flex grid to be able to make them
non-absolute, so I put them as a fixed part of TopBar instead. The value
is still fetched in the same place as before and passed upwards using
props, since gameEnabled is fetched in the same request and is needed by
Launch.svelte.

Fixes #183